% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_ad.tsets.spec}
\alias{estimate_ad.tsets.spec}
\alias{estimate_ad}
\title{Estimates an ETS model given a specification object using maximum likelihood and autodiff}
\usage{
\method{estimate_ad}{tsets.spec}(object, solver = "nlminb", control = list(trace = 0), ...)
}
\arguments{
\item{object}{An object of class tsets.spec.}

\item{solver}{One of either \dQuote{nloptr} or \dQuote{nlminb}. Using nlminb which takes
the hessian appears to provide better results.}

\item{control}{Solver control parameters.}

\item{...}{additional parameters passed to the estimation function}
}
\value{
An list of coefficients and other information.
}
\description{
Estimates an ETS model given a specification object using maximum likelihood and autodiff
}
\details{
This function is not expected to be used by itself but rather as a plugin
to be called from the estimate method of the tsets package.
}
