% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsetsad-package.R
\docType{package}
\name{tsetsad-package}
\alias{tsetsad}
\alias{tsetsad-package}
\title{tsetsad: ETS model estimation using automatic differentiation}
\description{
Plugin Estimator for the tsets package to use autodiff.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

}
\keyword{internal}
